Separate noisy linear as a file
